Update on the PR discussion
